requirements.txt
django==4.2.3
djangorestframework==3.14.0
django-extensions==3.2.3
django-db-connection-pool==1.2.5
openai==1.66.5
openapi-schema-pydantic==1.2.4
better-profanity==0.7.0
boto3==1.28.8
drf-spectacular==0.27.2
python-dotenv==1.0.1
serverless-wsgi==3.0.4
werkzeug==3.0.4
psycopg2-binary==2.9.9
psycopg2==2.9.9
pandas==2.2.2
openpyxl==3.1.5
langchain-postgres==0.0.9
SQLAlchemy==2.0.35
sqlparams==5.1.0
sqlparse==0.5.1
pypdf==5.0.0
unstructured==0.8.1
tiktoken==0.4.0
pydantic==2.11.0
pydantic-core==2.33.0
langsmith==0.1.96
langchain-community==0.2.11
langchain==0.2.12
python-docx==0.8.11
python-pptx==0.6.21
nltk==3.8.1
rank_bm25==0.2.2
scikit-learn==1.3.0
rich==13.9.4
openai-agents
pytesseract==0.3.13
pdf2image==1.17.0
pillow==11.1.0